+++ src/Gateway.php
@@ -59,6 +59,10 @@ class Pronamic_WP_Pay_Gateways_Mollie_Gateway extends Pronamic_WP_Pay_Gateway {
 				$request->method = Pronamic_WP_Pay_Mollie_Methods::MISTERCASH;
 
 				break;
+			case Pronamic_WP_Pay_PaymentMethods::SOFORT :
+				$request->method = Pronamic_WP_Pay_Mollie_Methods::SOFORT;
+
+				break;
 		}
 
 		$result = $this->client->create_payment( $request );