+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0
 *	Improved support for unsupported Mollie locales.
+*	Added support for the SOFORT Banking payment method.
 
 ## 1.0.0
 *	First release.